Move new change API to `stable`
